+++ core/src/main/java/pl/jutupe/core/repository/dao/GroupDao.kt
@@ -2,8 +2,6 @@ package pl.jutupe.core.repository.dao
 
 import androidx.paging.PagingSource
 import androidx.room.*
-import kotlinx.coroutines.flow.Flow
-import pl.jutupe.core.repository.dao.model.GroupWithDebtors
 import pl.jutupe.core.repository.entity.DebtorGroupCrossRef
 import pl.jutupe.core.repository.entity.Group
 
@@ -16,8 +14,8 @@ internal interface GroupDao {
     suspend fun insert(group: Group): Long
 
     @Transaction
-    suspend fun insertGroupWithDebtors(group: Group, debtorIds: List<Int>) {
-        val groupId = insert(group).toInt()
+    suspend fun insertGroupWithDebtors(group: Group, debtorIds: List<Long>) {
+        val groupId = insert(group)
 
         val refs = mutableListOf<DebtorGroupCrossRef>()
