+++ core/src/main/java/pl/jutupe/core/repository/entity/Debt.kt
@@ -6,7 +6,7 @@ import java.util.Date
 
 @Entity(tableName = "debts")
 data class Debt(
-    @PrimaryKey(autoGenerate = true) var debtId: Int,
+    @PrimaryKey(autoGenerate = true) var debtId: Long,
     var name: String,
     var amount: Int,
     var deadline: Date
